Refactor for design patten
